requirements.txt
aioredis==1.0.0
asgiref==2.2.0
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
certifi==2018.4.16
channels==2.1.5
channels-redis==2.0.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.3
Django==1.11.3
django-appconf==1.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-imagekit==4.0.2
django-redis==4.7.0
django-reversion==3.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
future==0.16.0
hiredis==0.3.0
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
msgpack-python==0.5.6
mysqlclient==1.3.13
pilkit==2.0
Pillow==5.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.23
uWSGI==2.0.17.1
XlsxWriter==1.0.5
xlwt==1.3.0
zope.interface==4.6.0
